LS: Move all hover logic into the `ide` module, remove dead code
commit-id:ff4f8cee
